Updated to iOS native SDK 2.3.2 and Android SDK 3.4.2
iOS
* Fixes issues with opened and received callbacks.
* Added additional notification_types status for "No Push Token" to help debug issues which will show on the OneSignal dashboard.
* Fixed rare crash with postNotification.
* Fixed issue opening a 2nd URL notification not showing a webview.
* Added validation to syncHashedEmail.
* Fixed issue where an iOS 10 media notification could hang the main thread when downloading it.

Android
* Added support for badges on older Samsung devices as well as OPPO an ZUK devices.
* Fixed SyncService logcat error.

+++ SDK/iOS/OneSignalIOS.cs
@@ -98,8 +98,8 @@ namespace Com.OneSignal
 
       public OneSignalIOS(string appId, bool autoPrompt, bool inAppLaunchURLs, OneSignal.OSInFocusDisplayOption displayOption, OneSignal.LOG_LEVEL logLevel, OneSignal.LOG_LEVEL visualLevel)
       {
-         var convertedLogLevel = (iOS.OneSLogLevel)((ulong)((int)logLevel));
-         var convertedVisualLevel = (iOS.OneSLogLevel)((ulong)((int)visualLevel));
+         var convertedLogLevel = (iOS.OneSLogLevel)((int)logLevel);
+         var convertedVisualLevel = (iOS.OneSLogLevel)((int)visualLevel);
 
          iOS.OneSignal.SetLogLevel(convertedLogLevel, convertedVisualLevel);
          var dict = new Foundation.NSDictionary("kOSSettingsKeyInAppLaunchURL", new Foundation.NSNumber(inAppLaunchURLs),
